% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_city.R
\name{select_city}
\alias{select_city}
\title{Select a city for shiny app}
\usage{
select_city(data)
}
\arguments{
\item{data}{A data frame containing a \code{city} variable}
}
\description{
Select a city which you want to display
}
\examples{
 
\dontrun{ 
select_city(cov_china)
}

}
